feat: apply UserArgumentResolver
- 사용자 인증 정보가 필요한 기존 컨트롤러 메소드에서 resolver를 통해 인증 정보 파라미터를 주입받도록 변경
- Issue #54

+++ src/main/java/com/team9ookie/dangdo/service/ReviewService.java
@@ -4,10 +4,7 @@ import com.team9ookie.dangdo.dto.file.FileDto;
 import com.team9ookie.dangdo.dto.file.FileType;
 import com.team9ookie.dangdo.dto.review.ReviewRequestDto;
 import com.team9ookie.dangdo.dto.review.ReviewResponseDto;
-import com.team9ookie.dangdo.entity.FileEntity;
-import com.team9ookie.dangdo.entity.Menu;
-import com.team9ookie.dangdo.entity.Review;
-import com.team9ookie.dangdo.entity.Store;
+import com.team9ookie.dangdo.entity.*;
 import com.team9ookie.dangdo.repository.FileRepository;
 import com.team9ookie.dangdo.repository.MenuRepository;
 import com.team9ookie.dangdo.repository.ReviewRepository;
@@ -83,13 +80,13 @@ public class ReviewService {
     }
 
     @Transactional
-    public long create(ReviewRequestDto dto) throws IOException {
+    public long create(ReviewRequestDto dto, User user) throws IOException {
         List<MultipartFile> fileList = dto.getFiles();
         long menuId = dto.getMenuId();
         Menu menu = menuRepository.findById(menuId)
                 .orElseThrow(() -> new IllegalArgumentException("메뉴를 찾을 수 없습니다. id: " + menuId));
         Store store = menu.getStore();
-        Review review = reviewRepository.save(dto.toEntity(store, menu));
+        Review review = reviewRepository.save(dto.toEntity(store, menu, user));
 
         List<FileEntity> fileEntityList = fileService.createFileList(fileList, FileType.REVIEW_IMAGE, review.getId());
         fileRepository.saveAll(fileEntityList);
@@ -98,9 +95,12 @@ public class ReviewService {
     }
 
     @Transactional
-    public long delete(long id) {
+    public long delete(long id, User user) {
         Review review = reviewRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("리뷰를 찾을 수 없습니다. id: " + id));
+        if (!review.getUser().getId().equals(user.getId())) {
+            throw new IllegalArgumentException("본인이 작성한 리뷰가 아닙니다.");
+        }
         reviewRepository.deleteById(review.getId());
         return review.getId();
     }
